+++ src/DotRecast.Detour.Extras/Jumplink/GroundSegment.cs
@@ -4,8 +4,8 @@ namespace DotRecast.Detour.Extras.Jumplink
 {
     public class GroundSegment
     {
-        public RcVec3f p = new RcVec3f();
-        public RcVec3f q = new RcVec3f();
+        public RcVec3f p;
+        public RcVec3f q;
         public GroundSample[] gsamples;
         public float height;
     }