+++ src/DotRecast.Detour.Extras/Jumplink/ITrajectory.cs
@@ -1,0 +1,10 @@
+using System;
+using DotRecast.Core.Numerics;
+
+namespace DotRecast.Detour.Extras.Jumplink
+{
+    public interface ITrajectory
+    {
+        public RcVec3f Apply(RcVec3f start, RcVec3f end, float u);
+    }
+}