+++ src/DotRecast.Detour.Extras/Jumplink/JumpTrajectory.cs
@@ -4,7 +4,7 @@ using DotRecast.Core.Numerics;
 
 namespace DotRecast.Detour.Extras.Jumplink
 {
-    public class JumpTrajectory : Trajectory
+    public class JumpTrajectory : ITrajectory
     {
         private readonly float jumpHeight;
 
@@ -13,7 +13,7 @@ namespace DotRecast.Detour.Extras.Jumplink
             this.jumpHeight = jumpHeight;
         }
 
-        public override RcVec3f Apply(RcVec3f start, RcVec3f end, float u)
+        public RcVec3f Apply(RcVec3f start, RcVec3f end, float u)
         {
             return new RcVec3f
             {